requirements/development.txt
# =====================================================================================
# DEVELOPMENT REQUIREMENTS - Development and testing dependencies
# =====================================================================================

# Include base requirements
-r base.txt

# Testing Framework
pytest==8.3.0
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-xdist==3.6.1  # Parallel test execution
factory-boy==3.3.1   # Test data factories
faker==28.4.1        # Fake data generation

# Code Quality & Linting
black==24.8.0        # Code formatter
flake8==7.1.1        # Linting
isort==5.13.2        # Import sorting
mypy==1.11.2         # Type checking
pylint==3.3.1        # Additional linting
pre-commit==3.8.0    # Git hooks

# Development Tools
ipython==8.27.0      # Enhanced Python shell
ipdb==0.13.13        # Debugger
django-extensions==3.2.3  # Django management extensions
django-debug-toolbar==4.4.6  # Debug toolbar
werkzeug==3.0.4      # WSGI utilities

# Documentation
sphinx==8.0.2        # Documentation generator
sphinx-rtd-theme==2.0.0  # ReadTheDocs theme
myst-parser==4.0.0   # Markdown support for Sphinx

# Monitoring & Profiling (Development)
flower==2.0.1        # Celery monitoring
django-silk==5.1.0   # Request/DB profiling
memory-profiler==0.61.0  # Memory profiling
line-profiler==4.1.3     # Line-by-line profiling

# Security Testing
bandit==1.7.9        # Security linter
safety==3.2.7        # Dependency security checker

# API Documentation
drf-spectacular==0.27.2  # OpenAPI 3.0 schema generation
drf-yasg==1.21.7     # Alternative Swagger generator

# Database Tools
django-querycount==0.8.3  # Query counting
nplusone==1.0.0      # N+1 query detection

# Hot Reload & Live Development
watchdog==5.0.2      # File system monitoring
django-browser-reload==1.15.0  # Browser auto-reload